fixes #45
the fulfillment API moved the `storePickupProductTitle` value to a child object, breaking InventoryWatch parsing. this value wasn't used by the app, though, so the fix was just to remove it as a parsing step. also bumps the version to v0.0.9.
