+++ InventoryWatch/Model/GithubTag.swift
@@ -1,0 +1,12 @@
+//
+//  GithubTag.swift
+//  InventoryWatch
+//
+//  Created by Worth Baker on 7/27/22.
+//
+
+import Foundation
+
+struct GithubTag: Codable {
+    let name: String
+}
