+++ InventoryWatch/Model/PartAvailability.swift
@@ -1,0 +1,23 @@
+//
+//  PartAvailability.swift
+//  InventoryWatch
+//
+//  Created by Worth Baker on 7/27/22.
+//
+
+import Foundation
+
+struct PartAvailability: Equatable, Hashable {
+    enum PickupAvailability: String {
+        case available, unavailable, ineligible
+    }
+    
+    let partNumber: String
+    let availability: PickupAvailability
+}
+
+extension PartAvailability: Identifiable {
+    var id: String {
+        partNumber
+    }
+}
